Librerias de seguridad terminadas
Pendiente implementacion en main
