--- storage/fase2/team15/storage/compresion.py
@@ -1,29 +1,0 @@
-from cryptography.fernet import Fernet
-
-
-def generarClave():
-    clave = Fernet.generate_key()
-    with open('clave.key', 'wb') as archivo:
-        archivo.write(clave)
-
-
-def cargar_clave():
-    return open('clave.key', 'rb').read()
-
-
-def encriptar(mensaje, llave):
-    f = Fernet(llave)
-    byteMessage = mensaje.encode()
-    return f.encrypt(byteMessage).decode()
-
-
-def desencriptar(mensaje, llave):
-    f = Fernet(llave)
-    return f.decrypt(mensaje.encode()).decode()
-
-
-clave = Fernet.generate_key()
-print(clave)
-encriptado = encriptar('dato1', clave)
-print(encriptado)
-print(desencriptar(encriptado, clave))
